+++ Plugin for WP sites/projects/rosenberger/theme/functions.php
@@ -87,7 +87,7 @@ add_filter(
 	'body_class',
 	function ( $classes ) {
 		if ( is_singular( 'property' ) || is_singular( 'reference' ) || is_post_type_archive( 'reference' )
-				|| is_home() || is_singular( 'post' ) ) {
+				|| is_home() || is_singular( 'post' ) || is_singular( 'region' ) ) {
 			$classes[] = 'has-light-hero';
 		}
 		return $classes;